bot/requirements.txt
discord.py
python-dotenv
dispander
lark-parser
motor
dnspython
